Move general-purpose util package into `x` folder (reminiscent of `golang.org/x/`).
