+++ src/rinq/amqp/peer.go
@@ -14,7 +14,6 @@ import (
 	"github.com/rinq/rinq-go/src/rinq/internal/opentr"
 	"github.com/rinq/rinq-go/src/rinq/internal/remotesession"
 	"github.com/rinq/rinq-go/src/rinq/internal/service"
-	"github.com/rinq/rinq-go/src/rinq/internal/syncutil"
 	"github.com/rinq/rinq-go/src/rinq/trace"
 	"github.com/streadway/amqp"
 )
@@ -194,11 +193,11 @@ func (p *peer) graceful() (service.State, error) {
 	p.remoteStore.GracefulStop()
 	p.listener.GracefulStop()
 
-	done := syncutil.Group(
-		p.remoteStore.Done(),
-		p.invoker.Done(),
-		p.server.Done(),
-		p.listener.Done(),
+	done := service.WaitAll(
+		p.remoteStore,
+		p.invoker,
+		p.server,
+		p.listener,
 	)
 
 	select {
@@ -223,11 +222,11 @@ func (p *peer) finalize(err error) error {
 		sess.Destroy()
 	})
 
-	<-syncutil.Group(
-		p.remoteStore.Done(),
-		p.invoker.Done(),
-		p.server.Done(),
-		p.listener.Done(),
+	<-service.WaitAll(
+		p.remoteStore,
+		p.invoker,
+		p.server,
+		p.listener,
 	)
 
 	closeErr := p.broker.Close()
