+++ src/rinq/internal/attrmeta/namespace.go
@@ -3,7 +3,7 @@ package attrmeta
 import (
 	"bytes"
 
-	"github.com/rinq/rinq-go/src/rinq/internal/bufferpool"
+	"github.com/rinq/rinq-go/src/rinq/internal/x/bufferpool"
 )
 
 // Namespace maps attribute keys to AttrRevs
