+++ src/rinq/internal/x/bufferpool/bufferpool_test.go
@@ -5,7 +5,7 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	. "github.com/rinq/rinq-go/src/rinq/internal/bufferpool"
+	. "github.com/rinq/rinq-go/src/rinq/internal/x/bufferpool"
 )
 
 var _ = Describe("Get", func() {
