+++ src/rinq/internal/x/repr/escape.go
@@ -1,4 +1,4 @@
-package strutil
+package repr
 
 import (
 	"encoding/json"
