+++ src/rinq/internal/x/repr/ginkgo_test.go
@@ -1,4 +1,4 @@
-package strutil_test
+package repr_test
 
 import (
 	"testing"
@@ -9,5 +9,5 @@ import (
 
 func TestSuite(t *testing.T) {
 	gomega.RegisterFailHandler(ginkgo.Fail)
-	ginkgo.RunSpecs(t, "strutil")
+	ginkgo.RunSpecs(t, "repr")
 }
