+++ src/rinq/internal/x/repr/pkg.go
@@ -1,0 +1,2 @@
+// Package repr contains utilities for building string representations of values.
+package repr
