+++ src/rinq/internal/x/syncx/lock.go
@@ -1,4 +1,4 @@
-package syncutil
+package syncx
 
 import "sync"
