+++ src/rinq/internal/x/syncx/pkg.go
@@ -1,0 +1,2 @@
+// Package syncx contains utilities for working with synchronization primitives.
+package syncx
