+++ src/rinq/payload.go
@@ -5,7 +5,7 @@ import (
 	"reflect"
 	"sync"
 
-	"github.com/rinq/rinq-go/src/rinq/internal/bufferpool"
+	"github.com/rinq/rinq-go/src/rinq/internal/x/bufferpool"
 	"github.com/ugorji/go/codec"
 )
